# DES/DES.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DES.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/encode": {
      "post": {
        "tags": [
          "DES"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Input64"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Input64"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Input64"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/api/encryptFile": {
      "post": {
        "tags": [
          "DES"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "Key": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              },
              "encoding": {
                "Files": {
                  "style": "form"
                },
                "Key": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OutputFile"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OutputFile"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OutputFile"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/decryptFile": {
      "post": {
        "tags": [
          "DES"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "Key": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              },
              "encoding": {
                "Files": {
                  "style": "form"
                },
                "Key": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OutputFile"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OutputFile"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OutputFile"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Input64": {
        "type": "object",
        "properties": {
          "input": {
            "type": "integer",
            "format": "int64"
          },
          "key": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "OutputFile": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}